ParticlesMaterial -> VFXMaterial
