+++ packages/vfx-composer/src/fiber/index.ts
@@ -1,4 +1,4 @@
 export * from "./Emitter"
 export * from "./makeParticles"
 export * from "./Particles"
-export * from "./ParticlesMaterial"
+export * from "./VFXMaterial"
